refactor(solver): extract TypePredicateCache impl to hold query_cache.rs under its size ratchet
The wave-5 rebase brought query_cache.rs to 2033 physical lines, over the
2022 size ratchet. Move the mechanical TypePredicateCache trait impl (every
method delegates to the interned TypeInterner caches) into a child module
caches/query_cache/predicate_cache.rs. Pure code motion: the child module
reads QueryCache's private interner field as a submodule, so no visibility
changes; behavior is unchanged. query_cache.rs 2033 -> 1861.
